requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdownify>=0.11.6
playwright>=1.40.0
pillow>=10.0.0
cairosvg>=2.7.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0
